move test to separate suite
